fix: creation product system
