+++ server/src/main/java/co/edu/uniquindio/comandera/infrastructure/spring/jpa/entity/entity/OrderProductEntity.java
@@ -1,5 +1,7 @@
 package co.edu.uniquindio.comandera.infrastructure.spring.jpa.entity.entity;
 
+import java.util.Set;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -8,6 +10,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import jakarta.persistence.Table;
 import jakarta.persistence.UniqueConstraint;
 
@@ -37,6 +40,9 @@ public class OrderProductEntity {
 
     @Column(nullable = true)
     private String image;
+
+    @OneToMany(mappedBy = "orderProduct")
+    private Set<OrderSubProductsEntity> subProducts;
     
     public OrderProductEntity() {
     }
@@ -111,5 +117,13 @@ public class OrderProductEntity {
 
     public void setImage(String image) {
         this.image = image;
+    }
+
+    public Set<OrderSubProductsEntity> getSubProducts() {
+        return subProducts;
+    }
+
+    public void setSubProducts(Set<OrderSubProductsEntity> subProducts) {
+        this.subProducts = subProducts;
     }
 }
